+++ src/main/java/de/markusbordihn/cats/component/CatTargetComponent.java
@@ -25,7 +25,7 @@ import com.hypixel.hytale.codec.codecs.UUIDBinaryCodec;
 import com.hypixel.hytale.component.Component;
 import com.hypixel.hytale.component.ComponentType;
 import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
-import de.markusbordihn.cats.Main;
+import de.markusbordihn.cats.Cats;
 import java.util.UUID;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -57,7 +57,7 @@ public class CatTargetComponent implements Component<EntityStore> {
 
   @Nullable
   public static ComponentType<EntityStore, CatTargetComponent> getComponentType() {
-    return Main.catTargetComponentType;
+    return Cats.catTargetComponentType;
   }
 
   @Nullable